+++ Source/Orts.Simulation/Common/Scripting/PowerSupply/AbstractPowerSupply.cs
@@ -56,7 +56,7 @@ namespace ORTS.Scripting.Api
         /// <summary>
         /// Called regularly at every simulator update cycle.
         /// </summary>
-        public abstract void Update(float elapsedClockSeconds);
+        public abstract void Update(double elapsedClockSeconds);
     }
 
     public enum PowerSupplyEvent
